api/requirements.txt
requests>=2.31
python-dotenv>=1.0
fastapi>=0.110
uvicorn[standard]>=0.27
google-genai>=0.3
pymongo[srv]>=4.6
google-auth>=2.28
google-auth-oauthlib>=1.2
google-api-python-client>=2.120
